[FEATURE] Support white/blacklist allowed FCE types
The same as the `allowedContentType` and `deniedContentTypes` feature in Flux:

```xml
<flux:grid.column variables="{Fluidcontent: {allowedContentTypes: 'fluidcontent_bootstrap:Alert.html'}}" />
```

And the same for `deniedContentTypes`. The setting is inherited from page column to container elements' columns, but not from container element to nested container element. Which means you can filter the allowed content types in both page template's Grid and FCE's content area.