[BUGFIX] Get rid of default file Index.html
The file is removed in order to let getTemplatePathAndFilename properly return NULL when no template file is resolved for a particular record. Doing so may cause some other places of code to fail (if returned template is not checked for NULL) but is a requirement for correctly processing records *not* belonging to fluidcontent.

Patch can be followed up with a feature that moves the "Fluid content type must be selected" message that is removed with this patch, to a proper system log or flash message.