[TASK] Handle non-ASCII hardcoded group names
This patch uses the md5() checksum of strings when the sanitised version of the string is empty, e.g. when it consists wholly of non-ASCII characters. Although it isn't the perfect fix (does not catch for example strings that become identical after sanitising; like `får` and `før` which both become `f_r` and thus uses the name of the *last* name used and groups elements from both those groups in the same tab) it improves on the current behavior by no longer causing "dead" groups/elements due to non-ASCII group names.